Revert "Update Unity's meta files"
This reverts commit 364ae2eefef2c1910956d80b7e587a472a6f6853.

Turns out that I mistakenly commit Unity meta files that were initially created on the first attempt to commit InControl as a submodule. During that first attempt it created these meta files and so on the second attempt, which is on the main project's branch `feature/controllerSupportViaInControlAtRootDirectory`, to completely remove and avoid creating these files. Unfortunately I failed to clean the directory of these files and inadvertently commit them again anyway. This commit hopefully rectifies this mistake and serves as warning to future developers.