Update Unity's meta files
Because this forked repository of InControl is now within another
project that parent project has updated its meta files to give them
updated GUIDs that relate it and not the subproject.

Also, because this directory has been moved underneath the Main
Project's Assets folder it needs to commit the required Unity meta
files that have generated for InControl's root directory.